tests(helpers): move some misc functions
